Add unitest for cufft
